FOOTPRINTING/DNS/requirements.txt
dnspython
argparse
python-tabulate